requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
openai==1.51.0
anthropic==0.36.2
google-generativeai==0.8.3
ollama==0.3.3
playwright==1.47.0
beautifulsoup4==4.12.3
lxml==5.3.0
openpyxl==3.1.5
pandas==2.2.3
duckduckgo-search==6.3.0
python-multipart==0.0.12
aiofiles==24.1.0
tenacity==9.0.0
